Refactor the configuration controller into multiple sub-controllers and move existing configuration classes into the new locations
