+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/controllers/configuration/security/ldap/support/LdapConstants.java
@@ -1,0 +1,24 @@
+package org.carlspring.strongbox.controllers.configuration.security.ldap.support;
+
+/**
+ * @author Przemyslaw Fusik
+ */
+public final class LdapConstants
+{
+
+    private LdapConstants()
+    {
+    }
+
+    public static class LdapMessages
+    {
+
+        public static final String NOT_CONFIGURED = "LDAP is not configured";
+
+        private LdapMessages()
+        {
+        }
+
+    }
+
+}
