--- strongbox-web-core/src/main/java/org/carlspring/strongbox/controllers/security/login/LoginInput.java
@@ -1,41 +1,0 @@
-package org.carlspring.strongbox.controllers.security.login;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-
-/**
- * @author Przemyslaw Fusik
- */
-@XmlRootElement(name = "login-input")
-@XmlAccessorType(XmlAccessType.NONE)
-public class LoginInput
-{
-
-    @XmlElement
-    private String username;
-
-    @XmlElement
-    private String password;
-
-    public String getUsername()
-    {
-        return username;
-    }
-
-    public void setUsername(String username)
-    {
-        this.username = username;
-    }
-
-    public String getPassword()
-    {
-        return password;
-    }
-
-    public void setPassword(String password)
-    {
-        this.password = password;
-    }
-}
