+++ strongbox-web-core/src/test/java/org/carlspring/strongbox/security/authentication/suppliers/CustomLoginSupplierTest.java
@@ -1,7 +1,7 @@
 package org.carlspring.strongbox.security.authentication.suppliers;
 
 import org.carlspring.strongbox.config.IntegrationTest;
-import org.carlspring.strongbox.controllers.security.login.LoginInput;
+import org.carlspring.strongbox.controllers.login.LoginInput;
 
 import javax.inject.Inject;
